Split ObjectSearchEnvironment from SymbolicEnvironment; add GreedyPlanner
SymbolicEnvironment is now domain-agnostic: no fluent or action name has
special meaning, and probabilistic branches sample from an instance-level
seedable RNG (seed=...) instead of the global random module. The
object-search conventions move to a new ObjectSearchEnvironment subclass:
ground-truth search resolution (true_object_locations), revelation
(searched -> revealed/found/at), robot intermediate {robot}_loc grounding,
and the move/place/search action hygiene filters. All search-domain users
(examples, procthor, unknown_search, replay) migrate to the subclass.

The greedy one-step-lookahead policy moves from pddl_converter.runner into
railroad.planner as GreedyPlanner, sharing the negative-precondition
preprocessing with MCTSPlanner instead of duplicating it.

Tests split accordingly: test_object_search_environment.py holds the
search-convention tests; test_symbolic_environment.py keeps the generic
loop/effect tests and pins the new generality (no revelation, no name
filtering, seeded reproducible sampling).
