+++ packages/railroad/src/railroad/environment/symbolic.py
@@ -171,7 +171,12 @@ class SymbolicEnvironment(Environment):
     Suitable for planning simulations and unit tests where:
     - Fluents are tracked in-memory
     - Skills execute by stepping through effects
-    - Probabilistic effects resolve via ground truth object locations
+    - Probabilistic effects are sampled with a seedable RNG
+
+    This class is deliberately domain-agnostic: no fluent or action name has
+    special meaning here. For railroad's object-search domains (ground-truth
+    search resolution, revelation of found objects, robot intermediate
+    locations), use :class:`ObjectSearchEnvironment`.
     """
 
     def __init__(
@@ -179,9 +184,9 @@ class SymbolicEnvironment(Environment):
         state: State,
         objects_by_type: Dict[str, Set[str]],
         operators: List[Operator] | None = None,
-        true_object_locations: Dict[str, Set[str]] | None = None,
         skill_overrides: Dict[str, Type[ActiveSkill]] | None = None,
         location_registry: LocationRegistry | None = None,
+        seed: int | None = None,
     ) -> None:
         """Initialize a symbolic environment.
 
@@ -190,26 +195,23 @@ class SymbolicEnvironment(Environment):
             objects_by_type: Objects organized by type.
             operators: Optional explicit operators for action instantiation.
                 If None, Environment.define_operators() is used.
-            true_object_locations: Ground truth object locations for search
-                resolution. If None, search always fails.
             skill_overrides: Optional mapping from action type prefix to skill
                 class. E.g., {"move": InterruptibleNavigationMoveSkill}
             location_registry: Optional LocationRegistry for interruptible moves.
                 Required by movement-path-aware skills such as
                 InterruptibleNavigationMoveSkill.
+            seed: Seed for the RNG used to sample probabilistic effect
+                branches. None seeds from entropy.
         """
         # Initialize subclass state before super().__init__
         # because _create_initial_effects_skill may need these
         self._fluents: Set[Fluent] = set(state.fluents)
         self._objects_by_type = {k: set(v) for k, v in objects_by_type.items()}
-        self._objects_at_locations = (
-            {k: set(v) for k, v in true_object_locations.items()}
-            if true_object_locations else {}
-        )
         self._skill_overrides: Dict[str, Callable[..., ActiveSkill]] = dict(
             skill_overrides or {}
         )
         self._location_registry = location_registry
+        self._rng = random.Random(seed)
 
         super().__init__(state=state, operators=operators)
 
@@ -341,37 +343,14 @@ class SymbolicEnvironment(Environment):
                     # Apply immediately and collect any further delayed effects
                     delayed_effects.extend(self.apply_effect(nested))
 
-        # Handle revelation (objects discovered when locations searched)
-        self._handle_revelation()
-
         return delayed_effects
-
-    def _handle_revelation(self) -> None:
-        """Reveal objects when locations are searched."""
-        for fluent in list(self._fluents):
-            if fluent.name == "searched":
-                location = fluent.args[0]
-                revealed_fluent = Fluent("revealed", location)
-
-                if revealed_fluent not in self._fluents:
-                    self._fluents.add(revealed_fluent)
-
-                    # Reveal objects at this location
-                    for obj in self._objects_at_locations.get(location, set()):
-                        self._fluents.add(Fluent("found", obj))
-                        self._fluents.add(Fluent("at", obj, location))
-                        self._objects_by_type.setdefault("object", set()).add(obj)
 
     def resolve_probabilistic_effect(
         self,
         effect: GroundedEffect,
         current_fluents: Set[Fluent],
     ) -> Tuple[List[GroundedEffect], Set[Fluent]]:
-        """Resolve probabilistic effect based on ground truth object locations.
-
-        For search actions, checks if target object is actually at location.
-        Otherwise, samples from the probability distribution.
-        """
+        """Sample one branch of a probabilistic effect with the seeded RNG."""
         if not effect.is_probabilistic:
             return [effect], current_fluents
 
@@ -379,68 +358,6 @@ class SymbolicEnvironment(Environment):
         if not branches:
             return [], current_fluents
 
-        # Find success branch (contains positive "found" fluent)
-        success_branch = None
-        target_object = None
-        location = None
-
-        for branch in branches:
-            _, branch_effects = branch
-            for eff in branch_effects:
-                for fluent in eff.resulting_fluents:
-                    if fluent.name == "found" and not fluent.negated:
-                        success_branch = branch
-                        target_object = fluent.args[0]
-                        location = self._find_search_location(eff, target_object)
-
-        # If we can resolve from ground truth, do so deterministically
-        if success_branch and target_object and location:
-            if self._is_object_at_location(target_object, location):
-                _, effects = success_branch
-                return list(effects), current_fluents
-            # Object not at location - return failure branch deterministically
-            other_branches = [b for b in branches if b is not success_branch]
-            if other_branches:
-                # No need to sample - ground truth determines the outcome
-                _, effects = other_branches[0]
-                return list(effects), current_fluents
-
-        # Can't determine from ground truth - sample from distribution
         probs = [p for p, _ in branches]
-        _, effects = random.choices(branches, weights=probs, k=1)[0]
+        _, effects = self._rng.choices(branches, weights=probs, k=1)[0]
         return list(effects), current_fluents
-
-    def _find_search_location(
-        self, effect: GroundedEffect, target_object: str
-    ) -> str | None:
-        """Find the location from 'at object location' fluent in a branch."""
-        for fluent in effect.resulting_fluents:
-            if fluent.name == "at" and len(fluent.args) >= 2:
-                if fluent.args[0] == target_object:
-                    return fluent.args[1]
-        return None
-
-    def _is_object_at_location(self, obj: str, location: str) -> bool:
-        """Check if object is at location using fluents + ground truth.
-
-        Priority:
-        1. If object is being held → not at any location
-        2. If fluent says object is at this location → yes
-        3. If fluent says object is at a different location → no
-        4. Fall back to ground truth (for undiscovered objects)
-        """
-        # Check if held by any robot
-        for f in self._fluents:
-            if f.name == "holding" and len(f.args) >= 2 and f.args[1] == obj:
-                return False
-
-        # Check fluents for known location
-        if Fluent("at", obj, location) in self._fluents:
-            return True
-
-        for f in self._fluents:
-            if f.name == "at" and len(f.args) >= 2 and f.args[0] == obj:
-                return False  # Object is at a different location
-
-        # Fall back to ground truth
-        return obj in self._objects_at_locations.get(location, set())
