+++ packages/railroad/src/railroad/examples/heterogeneous_robots.py
@@ -16,7 +16,7 @@ from railroad.core import Fluent as F, get_action_by_name
 from railroad.planner import MCTSPlanner
 from railroad.dashboard import PlannerDashboard
 from railroad import operators
-from railroad.environment import SymbolicEnvironment
+from railroad.environment import ObjectSearchEnvironment
 from railroad.environment.symbolic import (
     LocationRegistry,
 )
@@ -153,7 +153,7 @@ def main(
 
     # Initialize symbolic environment
     initial_state = State(0.0, initial_fluents, [])
-    env = SymbolicEnvironment(
+    env = ObjectSearchEnvironment(
         state=initial_state,
         objects_by_type=objects_by_type,
         operators=[no_op, move_op, search_op, pick_op, place_op],
